WIP UI Rewrite
- Moved the Github release checker code and program version
